server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiohttp-apispec==2.2.3
aiohttp-cors==0.8.1
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
apispec==3.3.2
asyncpg==0.30.0
attrs==25.3.0
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
ecdsa==0.19.1
flake8==7.2.0
Flask==3.1.1
flask-cors==6.0.0
frozenlist==1.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-genai==1.16.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.14.1
mccabe==0.7.0
multidict==6.4.4
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyflakes==3.3.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dotenv==1.1.0
python-jose==3.5.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
webargs==5.5.3
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.0
